added links in dashbord.js
