--- src/components/Buses.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import username from './images/username.png';
-import icon from './images/ICON.png'
-import add from './images/add.png';
-import route from './images/routes.png';
-import { RiLogoutBoxLine } from 'react-icons/ri';
-import { BiBus } from 'react-icons/bi';
-import { FaUsers } from 'react-icons/fa';
-import { BsPersonBadgeFill } from 'react-icons/bs';
-import { Link } from 'react-router-dom';
-
-function Buses() {
-  return (
-
-    <div>Buses</div>
-  )}
-
-export default Buses;
