+++ src/components/Dashboard/Buses.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+
+function Buses() {
+  return (
+
+    <div>Buses</div>
+  )}
+
+export default Buses;
